Prepare for the first release
* changed the group id and artifact id
* changed the main package based on the group id
* added all necessary plugins needed for deployment
