+++ src/main/java/co/igorski/clients/SamebugClient.java
@@ -1,0 +1,17 @@
+package co.igorski.clients;
+
+import co.igorski.model.CrashResponse;
+import co.igorski.model.SamebugRequest;
+
+import javax.ws.rs.Consumes;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.core.MediaType;
+
+@Path("/crashes")
+public interface SamebugClient {
+
+    @POST
+    @Consumes(MediaType.APPLICATION_JSON)
+    CrashResponse getCrashReport(SamebugRequest user);
+}
