+++ src/main/java/co/igorski/model/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains the model for the requests and responses to and from the Samebug api.
+ *
+ * @since 1.0.0
+ */
+package co.igorski.model;
