+++ src/main/java/co/igorski/services/ExceptionRequestFactory.java
@@ -1,0 +1,8 @@
+package co.igorski.services;
+
+import co.igorski.model.SamebugRequest;
+
+public interface ExceptionRequestFactory {
+
+    SamebugRequest createInstance(Throwable e);
+}
