--- src/main/java/org/igorski/model/CrashData.java
@@ -1,25 +1,0 @@
-package org.igorski.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "type",
-        "id"
-})
-@Getter
-@Setter
-@NoArgsConstructor
-public class CrashData {
-
-    @JsonProperty("type")
-    private String type;
-    @JsonProperty("id")
-    private Integer id;
-
-}
