--- src/main/java/org/igorski/model/CrashResponse.java
@@ -1,22 +1,0 @@
-package org.igorski.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "data"
-})
-@Getter
-@Setter
-@NoArgsConstructor
-public class CrashResponse {
-
-    @JsonProperty("data")
-    public CrashData data;
-
-}
