--- src/main/java/org/igorski/model/package-info.java
@@ -1,6 +1,0 @@
-/**
- * This package contains the model for the requests and responses to and from the Samebug api.
- *
- * @since 1.0.0
- */
-package org.igorski.model;
