--- src/main/java/org/igorski/services/ExceptionRequestFactory.java
@@ -1,8 +1,0 @@
-package org.igorski.services;
-
-import org.igorski.model.SamebugRequest;
-
-public interface ExceptionRequestFactory {
-
-    SamebugRequest createInstance(Throwable e);
-}
